Use uintptr_t for pointer arithmetic.
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2517 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/loader/HpSomBinaryFile.cpp
@@ -187,7 +187,7 @@ bool HpSomBinaryFile::RealLoad(const char* sName)
     unsigned* maxAux = auxHeaders + sizeAux;
     while (auxHeaders < maxAux)
         {
-            if ((UINT4(m_pImage + (int) auxHeaders) & 0xFFFF) == 0x0004)
+            if ((UINT4(m_pImage + (uintptr_t) auxHeaders) & 0xFFFF) == 0x0004)
                 {
                     found = true;
                     break;
@@ -238,7 +238,7 @@ bool HpSomBinaryFile::RealLoad(const char* sName)
 
     // A convenient macro for accessing the fields (0-11) of the auxilliary header
     // Fields 0, 1 are the header (flags, aux header type, and size)
-#define AUXHDR(idx) (UINT4(m_pImage + (int)(auxHeaders+idx)))
+#define AUXHDR(idx) (UINT4(m_pImage + (uintptr_t)(auxHeaders+idx)))
 
     // Section 0: header
     m_pSections[0].pSectionName = const_cast<char *> ("$HEADER$");
